requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
alembic
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0